+++ packages/ecount-dev-tool/src/stores/index.ts
@@ -1,4 +1,4 @@
 export * from './current_tab.svelte';
 export * from './accounts.svelte';
-export * from './section_collapse.svelte';
 export * from './section_visibility.svelte';
+export * from './section_order.svelte';